Promote a data structure to a member of Simulator
The IOInterfacePKG routine "constructCircuitFromNetlist" performs a
number of operations regarding the use of wild cards that would better
be performed much later.

One of the things it does requires access to an unordered set of
device names that was local to the function.  In order to move this
work elsewhere, I had to change the scope of this set.

Now, the set is a private member of the Simulator class, and a
reference to it is passed down into constructCircuitFromNetlist when
it is called by initializeEarly.

Later, in initializeLate I clear the names set so it doesn't waste
memory.

This commit does nothing but move the declaration of the set and make
it be passed into constructCircuitFromNetlist.  In future commits I'll
start moving the things that need this data structure out of
construcCircuitFromNetlist and into something that is called in
initializeLate.

The end goal is to make all the processing of wildcards in output
parameter lists happen in initializeLate and none of it in
constructCircuitFromNetlist.  This is so that external outputters can
function more completely as if they were print lines, instead of
making all this wildcard stuff happen ONLY for print lines.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29