Fix frequency domain loading to handle out of range frequencies
The freqquency domain loading the PULSE sources uses certain
number of samples. The overall frequency spectrum in the HB simulation
can have frequencies that are out of range of the sample frequencies
used in PULSE sources. This can lead to wrong results, non-convergence
and other wrong behaviors. This commit fixes the issue.

Xyce#219