Migrate detection of wildcard currents
When wildcarding support in current-based operators was improved to
include device lead currents instead of just device branch currents,
detection of the condition that "I(*)" had been put on print lines was
put into printLineDiagnostics, which is running through all the print
parameters given on .PRINT lines and doing some early error checking
and identification.

Unfortunately, putting this detection in this function means it is
happening too early to detect I(*) in external outputters (whose param
lists are not known until late initialization).

This commit simply rips the detection of those star variables into a
separate function and calls that function immediately after
printLineDiagnostics, with the expectation that later I'll be moving
the call elsewhere.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29