Move setting of instance wildcards
netlistImportTool was calling various functions to search output
parameter lists for device instances and wildcards and calling device
manager functions to set lead current requests.  To work on issue
243 (github 29) means I need this stuff to happen later.

This commit moves that stuff to where I think it needs to happen, late
enough to catch lead current requests in external outputters.

This is on the "Issue243_bad" branch because this is not in fact
working.  I do note that it seems to fix *part* of the issue, allowing
branch currents to be picked up properly by I(*) in an external
outputter, but does not enable lead currents in the wildcard unless
there is a corresponding print line with I(*) in it.

But worse, this stuff breaks a lot of test cases.  I believe the
primary reason it breaks is that there is an early determination of
linear/nonlinear devices, notably in the resistor.  This happens
before lead currents are finalized in this branch, which is busted.
I have to determine whether it is possible to defer that determination
in the resistor.  I think it may be, based on the fact that some other
devices make the determination at the first attempt to load the
matrices.

There is another spot in the device manager where it tries to
determine if the *entire* problem is linear, and this was also
happening too early.  I have moved that in this commit.  It isn't
enough to help.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29