Add input/output port names
The simulink interface requires the input
and output ports to be bundled in a mutex.
The order of the elements in the mutex is up
to the user and may not match Xyce's ordering.
Also, Xyce's ordering is not the same as it is
in the netlist because the DAC's and ADC's are
all accessed via maps.  So this commit adds
input and output parameters to the simulink
interface which specify names for the elements
in the I/O mutex.  This will let the interface
code get the same ordering as Xyce.  While that
isn't implemented yet, right now there is error
checking that the inputs and outputs as specified
by the user in Simulink match up to a DAC or ADC
from Xyce after Xyce has parsed the netlist.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104