Move resistor linear/nonlinear separation
The resistor was an early experiment in linear/nonlinear loading, and
had its code for making that determination happen very early, in the
addInstance call.

Later devices that do the linear/nonlinear separation do so much
later, upon the first call to the loader routines.

I have made the Resistor do the same thing, calling the Master class
function "separateInstances"  at the first call of loadDAEVectors or
loadDAEMatrices.  This breaks no tests.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29