Finish fixing issue 243 (github 29)
The final piece of the puzzle for why I(*) wasn't working with
external outputters should have been obvious much earlier, but there
were so many other things in the way it wouldn't be.

The function that scans output requests for the full I(*) (or any
wildcard that requires ALL lead currents to be computed) was scanning
only those output parameters in the output manager's
OutputParameterMap_.

Because External outputters are not created from print lines, an early
design choice I made was not to store them with full print_parameters
structures, since those have too many extra bits of data specific to
print lines.  Instead, pointers to the wrappers are stored in
ExternalOutputWrapperMap_, and the wrapper provides only the variable
list that would normally be a part of the print_parameters structure.

This means that every function that loops over OutputParameterMap_ to
get a list of output variables must *also* do an equivalent loop over
ExternalOuputWrapperMap_.  All of those functions *other* than the one
that looks for I(*) were doing that.

Now that we're doing wildcard resolution in late initialization
instead of at netlist parsing time (early initialization), this
matters, and fixing that one thing fixes what's left of issue 243.

This commit not only makes that change, but moves a function out
of the NetlistImportTool (where it's no longer called) and into the
one place where it *IS* called, N_CIR_Xyce.C.  It's still a free
function, but moving it makes more sense.

This commit also moves all the lead current checking stuff into a
separate member function of the simulator class, which is then called
as the first thing that initializeLate does.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29