Remove YLIN device's storeInstance method
The resistor device was an early attempt to separate linear and
nonlinear devices to be handled differently by the loader class.  In
this initial implementation, the segregation of devices was done at
device construction, in the storeInstance method.

Most other devices that can be either linear or nonlinear separate
their instances out much later, on the first call to either of the
master class load functions (loadDAEVectors or loadDAEMatrices).  In
an earlier commit I have already made the resistor like all the
others, which is necessary for resolution of issue 243.

The YLIN devices, which is strictly a frequency domain device at the
moment and is therefore linear, doesn't even use this segregation of
devices, but had a storeInstance method that was a stright copy of the
resistor's.

This commit simply removes that unnecessary method.

Should a nonlinear, time-domain version of the YLIN device ever be
needed, and lead currents enabled in that device, segregation of the
linear and nonlinear instances should be done the way all other
devices now do it.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29