Improve Error Handling for .FOUR and FOUR Measures
This commit improves the error handling for invalid (too small) FREQ
values on a .FOUR line and invalid AT values for a FOUR measure line.
Error checking of the FREQ value has been moved from the function
FourierMgr::getLastPeriod_() to FourierMgr::fixupFourierParameters().
This allows error checking for FREQ during netlist setup, rather than
causing an ungraceful exit at the end of the simulation. This is
similar to how error handling for other "dot lines" works.

A .MEASURE with a valid syntax but "nonsensical parameter values"
does not cause an error during netlist parsing.  Instead, the
measure is reported as "FAILED" and warnings may be emitted by
printMeasureWarnings() at the end of the simulation.  So, this
commit changes the Fourier::getMeasureResult() function to only
calculate the Fourier transform if a valid last period is found.
This allows the normal processing for Base::printMeasureWarnings()
to occur.  An additional function Fourier::printMeasureWarnings()
was added to deal with the fact that the AT parameter is a
frequency for the FOUR measure, rather than being a time value.
This split between those two printMeasureWarnings() functions is
NOT elegant, and might should be re-factored.

Issue #257
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/257