Partial Fix for Use of FOUR Measure on .PRINT TRAN
This commit is a partial fix for the use of a FOUR measure on a
.PRINT TRAN line. Prior to the related fix for issue #257, that
usage would cause Xyce to exit at the first time step when
Fourier::getMeasureResult() was called by the tran outputter.

This commit adds a new function, Fourier::calculateMeasureResult_(),
that is now only called at the end of the simulation.  This makes
the FOUR measure results output to stdout and the <netlist>.mtX
files, at the end of the simulation, correct even if a FOUR measure
is used on the .PRINT line.  This fix is incomplete, since the
FOUR measure values output to (for example) the <netlist>.prn
file are always the calculation default value, even if the
measure was not failed.

Git issues #255 and #257
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/255
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/257