Refactor detection of "I(*)" wildcards
This commit changes the function "processPrintParamIWildcards" into a
simple boolean function "isIStarRequested".  It is now a member
function of the output manager class instead of a free function, and
therefore requires no arguments.

I have refactored the code so that it makes use of the existing output
manager function "getVariableList" that aggregates the output variable
lists of all print lines and external outputters into a single
Util::ParamList container.  This enables isIStarRequested to be a
simple loop over all parameters, looking only for the star variables
of interest instead of a duplicated pair of loops over .print lines
and external outputters.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/258
Xyce#258