Removes unused sourceFlag from the DeviceBlock object
The DeviceBlock object stores a lot about a device during the parsing
and distribution phase, which is conveyed to the device manager during
allocation.  Some of the items stored in this object indicate which
type of device it is.  The "sourceFlag" that is in the DeviceBlock may
have indicated if the device was a current or voltage source at one time,
but it is not being used now.  So, it should be removed to avoid confusion
and save memory.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/250