Adds buffering to device block communication across processors
The addition of rebalancing of floating nodes can substantially
reduce the communication necessary for the simulator.  However,
if the number of devices being sent between processors is large,
then packing and unpacking these devices creates a bottleneck.
Buffering has been added to the rebalancing process to allow for
communication / computational overlap and provides a speedup.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/250