Fixes logic in .PREPROCESS ADDRESISTORS file output
There are several .PREPROCESS commands in Xyce that check for and
mitigate bad netlist syntax.  Some of these preprocessing commands
execute during the netlist parsing, but the ADDRESISTORS command
generates a whole new netlist will add resistors for voltage nodes
with no DC path and one terminal devices.

If the ADDRESISTORS command is used, the new netlist will include a
copy of the commented out .PREPROCESS command to serve as a reminder
that the netlist resulted from this preprocessing step.  Unfortunately,
the logic to do this is flawed and ALL of the .PREPROCESS commands
from the original netlist are commented out, even the ones that are
performed in situ (replace ground, remove redundant devices, etc.).

The logic is now fixed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/262