Exclude YGENEXT Devices from P() Wildcards
This commit refactors Xyce::IO:excludeYDeviceFromWildcard() into two
separate functions for I() and P() wildcards.  This is because
the YGENEXT device supports lead currents of the form IN, for N
equal to any positive integer, but it does not support power
calculations.  The code for extracting the Y device "basename"
is pulled out in a separate function, that is then used for both
I() and P() wildcards.

Issue #254
SON Bug 1317
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/254
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1317