Correct var type for internal variables.
Ages ago when I first worked on the nonlinear mutual inductor
two internal variables, M and R were given types of "M" and "R".
This didn't effect anything at the time, but is important now.
M, the magnetic moment, and R, the time derivative of the total
current flux, are both primarly functions of current and thus
should be of vartype "I" in Xyce.  This matters because if a user
tries the time integration option "maskivars" then the internal
variables M and R should be masked from the norm calculations.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/105