Improves OpenMP / ShyLU checks for CMake configuration
The CMake configuration now checks for the ShyLU_DDCore package in
ShyLU, which is the current subpackage that contains the ShyLU
hybrid-hybrid solver.  It also checks if OpenMP is enabled in Epetra
by looking for EPETRA_HAVE_OMP in the Epetra_config.h file.
This check has been added to the Kokkos device check to enable the
detection of OpenMP from Trilinos version 12.12 through current
development.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/182