Fix build for Windows/mingw32
The first fix is simply to add `-no-undefined` to the link line,
since Xyce is not using this feature and it is required for creating
shared libraries on windows.

The second fix is to remove the AC_F77_LIBRARY_LDFLAGS macro. This
macro is required when mixing C++ and Fortran *source files*, but
here we are merely linking a Fortran library, which works fine without
this macro. In theory the macro should not be harmful, but in practice
it is unfortunately broken. I am working on fixing that separately in
upstream autoconf, but for the moment, since it is not required here,
remove it to fix the build.

Lastly, the XYCESOURCES variable can apparently be empty, which will
cause libtool to attempt a C link rather than a C++ link. Manually
specifying C++ mode fixes this.