Apply patches for Windows/MINGW32 build
An open source user with the Julia project was trying to build Xyce
with Mingw ("Minimalist GNU for Windows") and found a few things in
our autoconf build that prevented it from workikng.  This is the
subject of github pull request GH-38, though the actual changes
present here are slightly different from what the user provided on his
pull request.

These fixes do not break any of our own use cases, but there is one
change that frightens me --- the removal of the call to
AC_F77_LIBRARY_LDFLAGS which forces our link line to link in Fortran
runtime libraries.  This is necessary when including Fortran and C++
source files in the same link (which we don't do), and *might* be
important if we try to link in static libraries full of Fortran
code (which we don't do, either, but which some users *might* try to
do).

This macro breaks the user's attempts to build with MINGW, though, and
its absence doesn't break any of our builds.  It should not break any
builds that include shared libraries of Fortran code (like most
system-provided BLAS and LAPACK libraries).

For the time being I'm commenting out the AC_F77_LIBRARY_LDFLAGS call,
but we may find users complaining that the build has suddenly broken
when they try to link against hand-rolled static BLAS and LAPACK
libraries.  Should that happen, we can come up with a scheme to allow
the macro to be called or not called based on a configure option.

The other changes tweak the link line and libtool options for building
libxyce.a (and libxyce.so).  They apparently allow this user to build
Xyce with shared library support on Windows with mingw, and do no harm
to our own builds.

Xyce#273
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/273
GH-38
https://github.com/Xyce/Xyce/pull/38