Refactor and Modify Output from .FFT Lines
This commit modifies the output generated by .FFT lines.  The header
info, for each FFT analysis in the <netlistName>.fft0 file, now
also contains the START and STOP times for each analysis.  In
addition, the values for all of the harmonics are now output in
that file.  The listings are no longer filtered, based on the FREQ,
FMIN and FMAX values specified for a given .FFT analysis.

This commit also adds the capability to also output the metrics (THD,
SNDR, ENOB, SNR and SFDR) for each FFT analysis to the stdout.  A
sorted list of the (up to) 30 largest harmonics can also be
sent to the stdout.  Both of these optional outputs are enabled
with .OPTIONS FFT FFTOUT=1.

SON Bug 1333
Issue #168
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1333
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/186