Add space to FFT analysis output
The FFT/StartStopTest test case has been failing on my system because
apparently the default justification of the stream is left justified,
and the ouptut line modified here has no space after a text field
prior to a numeric field.  On my system the columns run together
causing the test to fail because the field count on the line doesn't
match the gold standard.

Adding the space fixes my problem.  Hopefully it will not break
anywhere else.