Implement BINSIZ Parameter for .MEASURE FFT
This commit implements the BINSIZ parameter for .MEASURE FFT for
the SFDR, SNDR and ENOB measure types.  The BINSIZ parameter
allows the user to account for the "spectral broadening" of
the fundamental frequency used in those .MEASURE calculations.
The default BINSIZ value is 0.  In that case, the "signal energy"
is concentrated at the index (fhIdx_) specified by the FREQ
parameter (which has a default value of fhIdx_=1) on the associated
.FFT line.  For non-zero BINSIZ values, the "signal energy" is
assumed to be in the bins between (fhIdx_ - BINSIZ) and
(fhIdx_ + BINSIZ).  This is useful for modeling signals where
fhIdx_ is not equal to 1.  As a final note, this BINSIZ definition
may not match HSPICE.

Issue #256
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/256