Improve Handling of Measure Names
This commit improves the handling of measure names that start
with an "operator letter", such as D, I, N, P, S, V, W, Y or Z.
Those names (such as XMAX) now work when the measure is used as
the "measured quantity" in another measure, such as an EQN
measure.

It also fixes some corner-cases, for invalid .MEASURE lines,
as discussed in SON Bug 1134.

The final user-visible change is that the error messages,
generated by the new helper function extractOperatorData(),
have been modified to specified the operator type that caused
the error.

This commit moved the code for processing all operators that
are valid on .PRINT lines from extractPrintData() to a new helper
function extractOperatorData() that is now also usable by the
function extractMEASUREData().  This cleans up operator processing
during parsing of .MEASURE lines, without changing any of the other
existing code in extractPrintData() .  However, this refactoring
does have a few loose ends.  First, the "entrance criterion" to
extractOperatorData() is different for .PRINT lines vs. .MEASURE
lines.  For .PRINT lines, we look-ahead for a '(' character.  We
also do that for .MEASURE lines, but we also have to account for
the HSPICE PAR and ('<expString>') expression syntaxes.  Second,
extractPrintData() makes use of the error position and error
string info returned by extractOperatorData(), while
extractMEASUREData() does not.  Finally, the processing in the
function extractOperatorData() moves the "position" index (into
the parsed_line token vector) to the next position after the
closing ')' character of the operator.  For compatibility, with
the rest of extractMEASUREData() I then decrement that position
upon return from extractOperatorData().

Issues #277 and #278
SON Bug 1134
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/277
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1134