Add option to turn off BSIM3 SOI voltage limiting
Issue #282
"Make it so that BSIM3 SOI voltage limiting is optionally turned off"

It has been observed recently that for some large SOI circuits that the
version of voltage limiting implemented in the BSIM3 SOI device model
often does more harm than good, and leads to lots of failed time steps,
which would otherwise succeed.

The voltage limiting in the BSIM3 SOI device is an unorthodox
implementation, based on nodal voltages rather than junction voltages.
So, it has always seemed a bit suspect.  But, we implemented it that way
because that was what the BSIM group did in their Spice3
implementations.  Note, in the more recently Verilog-A implementations,
they didn't bother with it.

Note, in general, MOSFET devices are often robust without voltage
limiting, and there are a number of MOSFET device models in Xyce. This
includes the BSIM-SOI version 4, which came from a Verilog-A
implementation.

It has always been possible to globally turn off voltage limiting in
Xyce, via ".options device voltlim=0"  However, doing this
unconditionally, applied to ALL devices isn't the best solution to
this problem, as many CMOS circuit contain diodes, and diodes
really do need voltage limiting to be numerically robust.

So, for this issue I have implemented a new global device parameter

.options device B3SOIVOLTLIM=1/0

which allows the user to quickly disable voltage limiting in this device
model, while leaving it enabled in all other devices that might use it.

The default value for this flag is "1",  meaning it is "on" by default.