Remove redundant evaluateFunction call in capacitor
Issue #284
"Reduce redundant calls to "newExpression::evaluateFunction"
when using solution dependent capacitors"

This is a similar fix to one that I did for the Bsrc 3 months ago.  For
solution dependent capacitance or charge parameters in the capacitor
model, it was evaluating the expression two times.  Once, during the
standard "updateDependentParameters" function call, and then later
during the updatePrimaryState function in the capacitor model.

The first call was kind of automatic, but just provided the value of the
expression.  The second call was necessary to get the derivatives.  But,
the second call could provide both, so the first call was a waste of
resources.

Making this work correctly required a little more work than I originally
expected.  I had to make sure that the age and temperature models in the
capacitor were handled correctly.  They were being performed upstream
from the updatePrimaryState call, so I had to (basically) re-do these
operations in updatePrimaryState.

While doing that, I noticed that the capacitor derivatives were never
updated with respect to aging or temperature and they should have been.
So, this commit fixes that mistake.

Also, while doing that, I noticed that the capacitor had a mistake in
the logic of the constructor, where the expPtr was sometimes set
incorrectly.  The constructor loops over all the dependent parameters in
the capacitor.  At every stage of the loop, it sets expPtr = d->expr,
where "d" is the dependent parameter of the current iteration.  AFTER
setting this, it checks if the expression has numExpVars >0, or if it
is time dependent, and then proceeds (if true) to set up the extra
machinery for solution dependent capacitance(C) or charge(Q).  The problem
is, the loop doesn't terminate at this point and it keeps going.  And, if
there are additional dependent parameters, then expPtr will continue to
get updated until the loop is done.   This caused expPtr to be set to the
wrong pointer in the bug 1197 test case.  Instead of being set to the
expression pointer associated with the "C" parameter, it was set to the
expression pointer for the "temp" parameter.  So, instead of evaluating
to something like 4e-5, it was evaluating to temp+dtemp, which (after
conversion) was over 300.    This of course broke the value of "C"
completely, when I started getting "C" from the (wrong) expression.
It also meant that all the derivatives were (incorrectly) zero.

The constructor logic is also fixed in this commit, so that in cases
where there are multiple (correct) expression dependent parameters, the
expPtr variable is now set to the correct one.

Finally, I also updated one of the error messages in the capacitor to be
more informative.