Remove redundant evaluateFunction call from switch
Issue #286.
"Reduce redundant calls to "newExpression::evaluateFunction" when using
solution dependent switch devices"

I recently fixed the Bsrc and Capacitor in the same manner.  In all of
these devices, when there is a solution-dependent parameter, it is
declared to be a "dependent" parameter in the device package, and is
automatically updated during the "updateDependentParameters" function
call via the function newExpression::evaluateFunction.

However, for parameters like these, which depend on solution variables,
the device also needs derivatives.  So these devices each make a call to
newExpression::evaluate, which computes both the expression itself as
well as all its derivatives.  Since it evaluates the expression as well,
the original evaluateFunction call is a waste of resources.

In the switch device, it was possible to make this change by adding a
single line of code.  This commit implements that change.