Small bugfix, related to solution dependent caps
Issue #284
"Reduce redundant calls to "newExpression::evaluateFunction" when using solution dependent capacitors"

I fixed this issue the other day.  Doing so required fixing some broken
logic in the capacitor constructor.  Howvever, it was pointed out to me
today that I overlooked one issue in that fix.  This commit fixes that
overlooked issue.

The broken logic was causing the expPtr variable to be unconditionally
set at every pass thru a loop over dependent parameters.  The result was
that it was set to the expression pointer from the final parameter in
the vector, no matter which dependent parameter that happened to be.

This was fine if the final parameter in the vector was the "C" or "Q"
parameter, but was NOT fine if the final paremter was a non-C or non-Q
parameter.  expPtr is supposed to only point to the expression for C or
Q, and not anything else.    So, that got fixed the other day.

However, what I did NOT fix at the time was that the variable expNumVars
was also being set unconditionally.  This was a very similar bug, to
the above and I should have caught it.  But it is fixed now.