Improved Version of TRIG-TARG Measure
This commit adds new classes (TrigTargBase, TrigTarg and TrigTargCont)
that provide multiple improvements to the TRIG-TARG measures.
Highlights of the initial version of the added capabilities are:

  1) TRIG-TARG measures are now supported for AC, DC, NOISE and
TRAN measure modes.

  2) A TRIG-TARG measure will now correctly report a negative
measure value if the TRIG time comes after the TARG time.

  3) Separate TD values may now be specified for the TRIG and TARG
clauses.  (Note: If TD is only given for the TRIG clause then
that value is used for both the TRIG and TARG clauses.)

  4) The TARG clause may now use its own AT qualifier.

  5) TRAN_CONT, AC_CONT, DC_CONT, NOISE_CONT and TRAN_CONT
measure modes are now supported for TRIG-TARG measures.

Remaining open issues are:

  a) Better support for negative RISE/FALL/CROSS values.

  b) Ensuring that expressions work in all contexts for the
measured variable and target values in both the TRIG and
TARG clauses.

  c) Improved error handling in the checkMeasureLine() function.

  d) Improved error handling for failed measures.

Finally, for backward compatibility with Xyce 7.4 (mostly for
internal Sandia users), a new Xyce-specific .OPTIONS MEASURE
setting was added.  It is .OPTIONS MEASURE USE_LTTM=<value>,
where LTTM = "Legacy Trig Targ Mode".  The default value is 0
(or false).  However, the legacy RiseFallDelay class will still
be used (automatically in Manager::addMeasure) if the Xyce-specific
FRAC_MAX qualifier is used in either the TRIG or TARG clause.  This
LTTM setting may be a temporary fix, until the new TrigTarg classes
and the legacy RiseFallDelay classes can be integrated.  It is not
expected that external / open-source users would use the LTTM option.

Issues #220, #289, #302 and #303
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/220
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/303