Set R=0 resistor check to exclude more expressions
Issue #171
"Make Xyce directly support voltage-dependent resistors"

The device manager has a check in it to see if there are any R=0
resistors present, and if they are any, to replace them with 0V Vsrcs,
which are represented internally as "level 3 resistors".

This check doesn't work if the value of the resistor is not immutable.
So, some mutable expressions were being excluded from the check, but
only the mutable conditions that were allowed for resistors prior to my
recent changes for Issue #171.

The 3 conditions that would prevent expression-based resistors
from being subject to this test were isVariablesDependent,
isSpecialsDependent, and isRandomDependent.  This was insufficient, as
isVariablesDependent actually means "dependent on global parameters" and
not "solution variables" as one might expect.  (these are legacy names
from the old expression library).

The point of Issue #171 was to implement solution-dependent resistors
which I mostly accomplished in my previous commit today.  However,
solution dependent expressions (which depend on voltage nodes
and/or device currents in non-ohmic devices like Vsrc's) where not
checked for, so the relevant if-statement did not prevent them from
being subjected to R=0 checks.  But since a solution-dependent
epxression cannot produce an immutable value, the function was exciting
with an error.

So, the only way (prior to this commmit) that I could test out
solution-dependent resistors was to use the netlist option:

.options device CHECKFORZERORESISTANCE=false

This commit updates the function to check for more two more expression
conditions: isVoltageNodeDependent and isDeviceCurrentDependent.  Now, I
can run a solution-dependent resistor circuit and this check won't cause
a problem.

We should consider removing the automatic convertion to level=3
resistor, as I think there are functions in topology for removing R=0
resistors.  This topology functions arguably make this code obsolete.