Fix for ddx operator index mistake
Issue #290.
"Indexing mistake in the expression library's DDX operator"

I discovered this while working on Issue #171, which was about
implementing solution-dependent resistors.   I was creating tests for
that issue, and as part of those tests was using the ddx operator on the
.PRINT line, to help double check the derivatives.  I noticed that
sometimes these derivatives from DDX were wrong, when the
expression depended on multiple voltage nodes.

DDX operators are specified as {DDX(expr,var)}, where DDX is supposed to
compute d_expr/d_var.

In order for it to work, the ddx operator class has to find the node for
"var" in the expression somewhere.  So, the class has a function called
"resolveArg_" that is responsible for doing this.

The function first determines what type of argument "var" is.  param,
voltage or current.  Then once it determines this, it searches thru the
list of operators from the expression of that type.  Then, once it finds
it (if it finds it), it declares victory and stops.   I had a mistake in
the voltage op search section of this function.  It was a one-line
mistake that is now fixed.

As part of this commit I have added several new unit tests to test this
particular scenario.  While I had created many ddx tests, I was most
concerned with more complicated scenarios, such as propagating ddx thru
a table, or thru multiple layers of functions.  This unit test coverage
gap is now closed.