Small refinements to solution-dep resistor
Issue #171.
"Make Xyce directly support voltage-dependent resistors"

This commit moves the assignment of expNumVars = d->n_vars inside of the
relevant if-statement in the constructor.  It is the same small bugfix
that I yesterday applied to the capacitor.

This commmit also adds a (perhaps uncessary) line to zero out the
derivative array prior to calling "evaluate".