Add derivative corrections for temperature
Issue #171
"Make Xyce directly support voltage-dependent resistors"

I created a test for combining expressions with the resistor temperature
model.  While doing so I realized that I had made a mistake in the
setup.  The main conductance is updated with respect to temperature, but
the derivatives that come from the expression library were not being
updated.  This commit corrects that mistake.

It similar to a mistake that was present in the capacitor, which I fixed
previously.  So, I should have caught it, but at least it is fixed now.