Additional fix for period separators
Issue #223
"Add a new subcircuit separator, "." instead of ":" as part of -hspice-ext"

It was pointed out to me today that the new period separator capability
didn't work in at least one use case, that of printing "interface nodes"
on the .print line.  This was an oversight, and the fix involved
changing a single line of code.  This commit contains that one-line
change.