Improve DERIV measure behavior at first simulation step
This commit updates DerivativeEvaluationBase::isWHENcondition()
DerivativeEvaluationBase::interpolateCalculationInstant() to be
the same as the corresponding functions in the FindWhenBase class.
This fixes some issues with the behavior of the DERIV-WHEN
measure at the first simulation step for AC, DC, NOISE and TRAN
measure modes.  The longer-term solution is to refactor the classes
for the DERIV, FIND and WHEN measures to use a common base class.

Issue #291
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/291