Remove unconditional debugging output
In commit 3e164d0, @hkthorn fixed a bug in the ADDRES preprocessing
that failed to break out of a loop when it should have, causing all
subsequent iterations to comment things out that it shouldn't have.

In doing this fix, Heidi had put some debugging output into the code
without wrapping it in debugging if(DEBUG_IO) statements (a reasonable
thing to do for debugging output you intend to remove after the
hacking run, but not for stuff that stays in the code).  She forgot to
remove it.

I'm removing it as part of my QA.

Xyce#262
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/262