Fix for soln-dependent resistors in subcircuits
Issue #171
"Make Xyce directly support voltage-dependent resistors"

The newly implemented solution dependent resistors did not work
inside of subcircuits, at least for some use cases.  This was pointed
out to me by Tom when he reviewed my work on this issue.

Specifically, if you have a solution dependent resistor inside of a
subcircuit, and the expression depends on a node that is local to that
subcircuit, the names cannot be resolved unless the subcircuit prefix is
added to the node name.    There is code in the parser that handles
this, but it only does it for cases where solution-dependent parameters
are allowed, such as for Bsrcs, capacitors and (now) resistors.

With this commit, this oversight is fixed.