Refactor DERIV, FIND and WHEN measures to use WhenAT Class
This commit refactors the classes for the DERIV-AT, DERIV-WHEN,
FIND-AT, FIND-WHEN and WHEN measures to use a common WhenAT
base class.  This consolidates a lot of common code between
those classes. It was motivated by issue #291, which occured
because changes to the functions FindWhenBase::isWHENcondition()
and FindWhenBase::interpolateCalculationInstant() were not mirrored
in the corresponding functions in DerivativeEvaluationBase.
This commit also cleans up some public vs. protected vs. private
access designators for various functions, in the classes changed
by this commit, which is a follow-on to issue #292.

Git issue #293
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/293